complete DataAcces layer with entity framework and Migrations

+++ Sat.Recruitment.Model/User.cs
@@ -10,6 +10,6 @@ namespace Sat.Recruitment.Model
         public string? Address { get; set; }
         public string? Phone { get; set; }
         public string? UserType { get; set; }
-        public decimal Money { get; set; }
+        public double  Money { get; set; }
     }
 }
